backend/requirements.txt
aiosqlite==0.20.*
alembic==1.13.*
Authlib==1.3.0
fastapi==0.111.*
httpx==0.27.*
ipython==8.26.0
itsdangerous==2.2.0
Jinja2==3.1.*
asyncpg==0.29.*
pytest==8.3.*
requests==2.32.*
SQLAlchemy==2.0.*
uvicorn==0.30.*
bcrypt==4.2.*
sqlalchemy-utils==0.41.*
python-multipart==0.0.*
pyjwt==2.9.*
python-dateutil==2.9.*
pydantic-settings==2.4.*
pydantic[email]==2.8.*
typing-extensions==4.8.0
python-socketio==5.11.4
lxml==4.9.3